pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify>=2.1.0",
    "apify-client>=1.8.1",
    "boto3>=1.35.76",
    "pymilvus",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "pillow>=11.0.0",
]
